override "to_s" method in "assert_block" method
